pyproject.toml
[project]
name = "symphony-py"
version = "0.3.0"
description = "Python Symphony-style orchestrator for Codex App Server"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.8",
  "httpx>=0.27",
  "PyYAML>=6.0",
  "fastapi>=0.115",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.6",
  "mypy>=1.11",
]

[project.scripts]
symphony-py = "symphony_py.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = false
